+++ src/ch.usi.inf.nodeprof.test/src/ch/usi/inf/nodeprof/test/examples/NonContiguousArray.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright [2018] [Haiyang Sun, Università della Svizzera Italiana (USI)]
+ * Copyright 2018 Dynamic Analysis Group, Università della Svizzera Italiana (USI)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -54,7 +54,7 @@ public class NonContiguousArray extends TestableNodeProfAnalysis {
     public void printResult() {
         int idx = 0;
         Logger.info("NonContiguousArray analysis finishes.");
-        for (Entry<Long, Report> entry : db.getRecords().entrySet()) {
+        for (Entry<Integer, Report> entry : db.getRecords().entrySet()) {
             Logger.info("NonContiguousArray [" + idx++ + "]" + entry.getValue().report());
         }
     }
